saving location now accepts googleId in params
