+++ api/routes/savedLocationsRoute.js
@@ -1,15 +1,17 @@
-const SAVED = require("../models/SavedLocationsModel");
 const router = require("express").Router();
+const SAVED = require("../models/SavedLocationsModel");
+
 const authenticate = require("../middleware/authenticate.js");
+const findLocation = require("../middleware/locations/findLocation");
 
 router.use(authenticate);
 
 // @route  /locations/save/:locationId
 // @desc   Add a location to users saved locations
 // @access Basic Users
-router.post("/:id", async (req, res) => {
-  const userId = res.locals.decodesJwt.userId;
-  const locationId = req.params.id;
+router.post("/:id", findLocation, async (req, res) => {
+  const userId = res.locals.decodedToken.userId;
+  const locationId = res.locals.location.id;
   const success = await SAVED.addSavedLocation(userId, locationId);
   !!success
     ? res.status(201).json({ message: "Location saved." })
